Rename hash.h to hash.hpp and cleanup includes.
