+++ src/tests/key-add-userid.cpp
@@ -28,7 +28,6 @@
 #include "pgp-key.h"
 #include "rnp_tests.h"
 #include "support.h"
-#include "crypto/hash.h"
 
 /* This test loads a pgp keyring and adds a few userids to the key.
  */
